refactor(docsv/scripts): extract shared doc utilities into doc_utils.py
Resolves Cubic P2: params_table_rows / _params_table were near-identical
across the two generator scripts. The duplication was wider than just the
table function — clean_rst_markup, parse_param_list, parse_google_docstring,
and format_type_hint were all copied between the files.

Move all five shared utilities to a new doc_utils.py module:
- clean_rst_markup: use the fuller internals version (handles
  .. code-block::, .. note::, .. warning::, double backticks, :ref:,
  :py:..., :code: — a superset of what api-docs had)
- parse_param_list, parse_google_docstring: identical, now single copy
- format_type_hint: internals version uses a broader sqlfluff path regex
- params_table_rows: unified name (was _params_table in internals)

Both scripts now import from doc_utils; re, textwrap, and the duplicate
function bodies are removed from each.
