+++ docsv/scripts/generate-internals-docs.py
@@ -8,10 +8,16 @@ Markdown pages + a VitePress sidebar JSON to reference/internals/.
 
 import inspect
 import json
-import re
-import textwrap
 from pathlib import Path
 from typing import Any, Callable, get_type_hints
+
+from doc_utils import (
+    format_type_hint,
+    parse_google_docstring,
+)
+from doc_utils import (
+    params_table_rows as _params_table,
+)
 
 from sqlfluff.core.config import loader as config_loader
 from sqlfluff.core.config.fluffconfig import FluffConfig
@@ -110,149 +116,6 @@ Source:
 """,
 }
 
-# ── docstring parsing (shared with generate-api-docs.py) ─────────────────────
-
-
-def clean_rst_markup(text: str) -> str:
-    """Remove RST-style markup from text and convert to Markdown equivalents."""
-    if not text:
-        return text
-    # Cross-reference roles
-    text = re.sub(r":obj:`([^`]+)`", r"`\1`", text)
-    text = re.sub(r":class:`([^`]+)`", r"`\1`", text)
-    text = re.sub(r":func:`([^`]+)`", r"`\1`", text)
-    text = re.sub(r":meth:`([^`]+)`", r"`\1`", text)
-    text = re.sub(r":attr:`([^`]+)`", r"`\1`", text)
-    text = re.sub(r":mod:`([^`]+)`", r"`\1`", text)
-    text = re.sub(r":py:[a-z]+:`([^`]+)`", r"`\1`", text)
-    text = re.sub(r":ref:`([^`]+)`", r"\1", text)
-    text = re.sub(r":code:`([^`]+)`", r"`\1`", text)
-
-    # RST code blocks  →  fenced Markdown
-    def _code_block(m: re.Match) -> str:
-        lang = (m.group(1) or "").strip()
-        lang = {
-            "py3": "python",
-            "py": "python",
-            "cfg": "ini",
-            "bash": "bash",
-            "sql": "sql",
-            "text": "",
-        }.get(lang, lang)
-        # Extract indented body
-        body_lines = []
-        for line in m.group(2).splitlines():
-            body_lines.append(line[4:] if line.startswith("    ") else line)
-        return f"```{lang}\n" + "\n".join(body_lines).strip() + "\n```"
-
-    text = re.sub(
-        r"\.\.\s+code-block::\s*(\w*)\n\n((?:    .+\n?|\n)+)",
-        _code_block,
-        text,
-    )
-    # RST ``..`` directives: note/warning → VitePress containers
-    text = re.sub(
-        r"\.\.\s+note::\s*\n\n((?:    .+\n?|\n)+)",
-        lambda m: "::: info\n" + textwrap.dedent(m.group(1)).strip() + "\n:::\n",
-        text,
-    )
-    text = re.sub(
-        r"\.\.\s+warning::\s*\n\n((?:    .+\n?|\n)+)",
-        lambda m: "::: warning\n" + textwrap.dedent(m.group(1)).strip() + "\n:::\n",
-        text,
-    )
-    # Leftover bare ``.. something::`` directives — strip them
-    text = re.sub(r"\.\.\s+\w+::[^\n]*\n", "", text)
-    # RST ``double backtick`` → single backtick
-    text = re.sub(r"``([^`]+)``", r"`\1`", text)
-    return text
-
-
-def parse_param_list(text: str) -> list[dict[str, str]]:
-    """Parse Google-style parameter list into structured dicts."""
-    if not text.strip():
-        return []
-    params: list[dict[str, str]] = []
-    param_pattern = r"^\s*(\w+)\s*(?:\(([^)]+)\))?\s*:\s*(.+)$"
-    current: dict[str, str] | None = None
-    for line in text.split("\n"):
-        m = re.match(param_pattern, line)
-        if m:
-            if current:
-                params.append(current)
-            current = {
-                "name": m.group(1),
-                "type": clean_rst_markup(m.group(2) or ""),
-                "description": clean_rst_markup(m.group(3).strip()),
-            }
-        elif current and line.strip():
-            current["description"] += " " + clean_rst_markup(line.strip())
-    if current:
-        params.append(current)
-    return params
-
-
-def parse_google_docstring(docstring: str) -> dict[str, Any]:
-    """Parse a Google-style docstring into keyed sections."""
-    if not docstring:
-        return {}
-    cleaned = inspect.cleandoc(docstring)
-    sections: dict[str, Any] = {
-        "description": "",
-        "args": [],
-        "returns": "",
-        "raises": [],
-        "examples": "",
-        "note": "",
-    }
-    section_pattern = r"^(Args?|Returns?|Raises?|Examples?|Note):\s*$"
-    lines = cleaned.split("\n")
-    current_section = "description"
-    current_content: list[str] = []
-
-    def _flush(section: str, content: list[str]) -> None:
-        text = "\n".join(content).strip()
-        if section == "description":
-            sections["description"] = clean_rst_markup(text)
-        elif section in ("args", "raises"):
-            sections[section] = parse_param_list(text)
-        else:
-            sections[section] = clean_rst_markup(text)
-
-    for line in lines:
-        m = re.match(section_pattern, line.strip())
-        if m:
-            _flush(current_section, current_content)
-            key = m.group(1).lower().rstrip("s")
-            current_section = {
-                "arg": "args",
-                "return": "returns",
-                "raise": "raises",
-                "example": "examples",
-            }.get(key, key)
-            current_content = []
-        else:
-            current_content.append(line)
-    _flush(current_section, current_content)
-    return sections
-
-
-# ── type-hint formatting ──────────────────────────────────────────────────────
-
-
-def format_type_hint(hint: Any) -> str:
-    """Format a type hint for Markdown display."""
-    if hint is None or hint is type(None):
-        return "None"
-    if isinstance(hint, str):
-        return hint
-    s = str(hint)
-    s = s.replace("typing.", "")
-    s = s.replace("<class '", "").replace("'>", "")
-    s = re.sub(r"sqlfluff\.\w+(?:\.\w+)*\.", "", s)
-    return s
-
-
 # ── introspection helpers ─────────────────────────────────────────────────────
 
 
@@ -329,30 +192,6 @@ def _return_type(obj: Any) -> str:
 
 
 # ── markdown rendering ────────────────────────────────────────────────────────
-
-
-def _params_table(rows: list[dict]) -> list[str]:
-    """Render a 4-column parameter table wrapped in a .params-table div.
-
-    The div enables CSS to apply min-widths on wide viewports and a stacked
-    definition-list layout on narrow ones, without touching other tables.
-    """
-    if not rows:
-        return []
-    lines = [
-        "**Parameters:**\n",
-        '<div class="params-table">\n',
-        "| Parameter | Type | Default | Description |",
-        "|:----------|:-----|:--------|:------------|",
-    ]
-    for r in rows:
-        name = f"`{r['name']}`"
-        typ = f"`{r['type']}`" if r.get("type") else " "
-        default = f"`{r['default']}`" if r.get("default") else " "
-        desc = r.get("description", "").replace("\n", " ").strip() or " "
-        lines.append(f"| {name} | {typ} | {default} | {desc} |")
-    lines.append("\n</div>\n")
-    return lines
 
 
 def render_function(name: str, obj: Callable, heading: str = "###") -> list[str]:
